Temporary hack to make the example story compile
The compiler does not yet properly implement inheritance. Therefore
the example story, which defines a base object and two child objects
that use a member inherited from it, does not compile.

To work around this, these objects redefine the member for
now. Obviously this is not the correct thing to do and should be
undone once inheritance is supported.