Re-add support for nested objects
Instead of hard-coding the object that a variable belongs to, which
was the previous, broken solution, now each object keeps track of its
enclosing object. When accessing a member of an outer object, we just
use that reference, which will work fine with inheritance as then the
nested object inside the child will just have a reference to that
child rather than the parent.